vstrt: implement flexible output
The flexible output mode (enabled with "flexible_output_prop") can return arbitrary number of output planes.
In the initialization stage of the filter instance, the output is augmented with the number of output planes.
The actual output frames are stored in the property map of the return frame.

Therefore,

  prop = "planes"
  output = core.trt.Model(src, engine_path, flexible_output_prop=prop)
  planes = [core.std.PropToClip(output["clip"], prop=f"{prop}{i}") for i in range(output["num_planes"])]

will return a list of output planes.